+++ examples/angular/src/app/api/api.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-api',
+  templateUrl: './api.component.html',
+})
+export class ApiComponent {
+  descriptionUrl = 'https://raw.githubusercontent.com/stoplightio/studio-demo/master/reference/todos/openapi.v1.json';
+}