+++ examples/angular/src/app/docs/docs.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-docs',
+  templateUrl: './docs.component.html',
+})
+export class DocsComponent {
+  workspace = 'https://elements.stoplight.io';
+  project = 'studio-demo';
+}