+++ packages/elements-web-components/src/__stories__/index.ts
@@ -1,1 +1,2 @@
-export * from './SchemaViewer';
+export * from './Api';
+export * from './StoplightProject';