+++ packages/elements/src/__stories__/index.ts
@@ -2,3 +2,5 @@ import './components/Dependencies';
 import './components/Docs';
 import './components/TableOfContents';
 import './components/TryIt';
+import './components/Api';
+import './components/StoplightProject';