+++ packages/elements/src/components/RequestMaker/Request/Mocking.tsx
@@ -3,7 +3,7 @@ import cn from 'classnames';
 import { observer } from 'mobx-react-lite';
 import * as React from 'react';
 
-import { useRequestMakerStore } from '../../../hooks/useRequestMaker';
+import { useRequestMakerStore } from '../../../hooks/useRequestMakerStore';
 
 const notSetOption = { value: '', label: 'Not Set' };