+++ packages/elements/src/components/RequestMaker/SuggestionBar.tsx
@@ -1,7 +1,7 @@
 import { observer } from 'mobx-react-lite';
 import * as React from 'react';
 
-import { useRequestMakerStore } from '../../hooks/useRequestMaker';
+import { useRequestMakerStore } from '../../hooks/useRequestMakerStore';
 import { SuggestionFunc, useSuggestion } from '../../hooks/useSuggestion';
 import { RequestMakerStore } from '../../stores/request-maker';