+++ packages/elements/src/index.ts
@@ -1,6 +1,4 @@
-export * from './containers/Dependencies';
-export * from './containers/Docs';
-export * from './containers/Provider';
-export * from './containers/TableOfContents';
-export * from './containers/TryIt';
-export { SchemaViewer, ISchemaViewerProps } from './components/SchemaViewer';
+export { StoplightProject } from './containers/StoplightProject';
+export { API } from './containers/API';
+
+export { IAPI, IStoplightProject } from './types';